+++ examples/data-sources/graphiant_content_filters/data-source.tf
@@ -1,0 +1,5 @@
+data "graphiant_content_filters" "all" {}
+
+output "content_filter_names" {
+  value = [for f in data.graphiant_content_filters.all.content_filters : f.name]
+}